requirements.txt
APScheduler==3.9.1.post1
flake8==5.0.4
loguru==0.6.0
mccabe==0.7.0
pyarmor==7.7.3
pyminifier==2.1
radon==5.1.0
rsa==4.9
pycryptodome==3.15.0
pandas==1.5.2